refactor(client): S3 Cursor+Kimi feature-first provider families
Move Cursor and Kimi owners into providers/{cursor,kimi-code}, split the mixed
discover-models Runtime owner into family modules, and drain transitional
capability/binary/login paths from 14 to 8 while keeping public export identity.
