+++ packages/client/src/__tests__/provider-support-export-allowlists.ts
@@ -228,17 +228,11 @@ export const PROVIDER_SUPPORT_EXPORT_ALLOWLISTS = {
 } as const satisfies Record<string, readonly string[]>;
 
 export const TRANSITIONAL_PROVIDER_FAMILY_FILES = [
-  "runtime/capabilities/cursor.ts",
   "runtime/capabilities/detect.ts",
-  "runtime/capabilities/discover-models.ts",
   "runtime/capabilities/index.ts",
-  "runtime/capabilities/kimi-code.ts",
   "runtime/capabilities/launch-probe.ts",
   "runtime/capabilities/opencode.ts",
   "runtime/capabilities/pi.ts",
-  "runtime/cursor-binary.ts",
-  "runtime/cursor-login.ts",
-  "runtime/kimi-binary.ts",
   "runtime/opencode-binary.ts",
   "runtime/opencode-private-config.ts",
   "runtime/pi-binary.ts",
