+++ packages/client/src/providers/auth-drivers.ts
@@ -1,8 +1,8 @@
 import type { RuntimeAuthProvider } from "@first-tree/shared";
-import { createCursorAuthDriver } from "../runtime/cursor-login.js";
 import type { RuntimeAuthDriver } from "./auth-driver.js";
 import { createClaudeAuthDriver } from "./claude/login.js";
 import { createCodexAuthDriver } from "./codex/login.js";
+import { createCursorAuthDriver } from "./cursor/login.js";
 import { createGrokAuthDriver } from "./grok/login.js";
 
 /**
