+++ packages/client/src/providers/cursor/__tests__/fixtures/cursor/multitool-interleaved.stream.jsonl
