+++ packages/client/src/providers/cursor/__tests__/fixtures/cursor/quota-failure.stream.jsonl
