+++ packages/client/src/providers/cursor/__tests__/fixtures/cursor/shell-exit7-turn-success.stream.jsonl
