+++ packages/client/src/providers/cursor/__tests__/fixtures/cursor/success-connectivity.stream.jsonl
