+++ packages/client/src/providers/cursor/parser.ts
