+++ packages/client/src/providers/kimi-code/__tests__/handler.test.ts
@@ -11,11 +11,11 @@ import type {
 } from "@botiverse/kimi-code-sdk";
 import type { AgentRuntimeConfigPayload, SessionEvent } from "@first-tree/shared";
 import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
-import { createKimiCodeHandler, formatKimiCodeError, kimiToolIsReadOnly } from "../handlers/kimi-code.js";
-import type { DeliveryToken, SessionContext, SessionMessage, TurnOutcome } from "../runtime/handler.js";
-import { noopDeliveryToken } from "../runtime/handler.js";
-import type { ReplayFenceEntry } from "../runtime/replay-fence.js";
-import { mockCtxPlumbing } from "./test-helpers.js";
+import { mockCtxPlumbing } from "../../../__tests__/test-helpers.js";
+import type { DeliveryToken, SessionContext, SessionMessage, TurnOutcome } from "../../../runtime/handler.js";
+import { noopDeliveryToken } from "../../../runtime/handler.js";
+import type { ReplayFenceEntry } from "../../../runtime/replay-fence.js";
+import { createKimiCodeHandler, formatKimiCodeError, kimiToolIsReadOnly } from "../index.js";
 
 class FakeSession {
   readonly id: string;
